rename sparqle.ts to sparql.ts
